ADD API FUNCTIONALITY BASED ON ACCESS/REFRESH TOKENS SENT AS HTTP COOKIES
Set up proxy for API in the vue.config file to proxy '/api' requests to
the API server. This allows the webpack server to contact the api but
keeps the same url in the browser - this overcomes CORS issues with
complex requests being blocked and http cookies not being set in the
browser of they originate from a different origin.

Added async actions in the namespaced modules to perform REST
Api requests and set fetch api to send credentials to auto send
http cookies with each request for authorization.

Set up a non-http cookie which was set at the same time as the short-lived
Access token Cookie - this allows Vue to check if the cookie is set before
eahc fectch request in case we need to send a refresh request with the refresh
token cookie to get a new Access Token cookie - this prevents 401 errors.

Login and Register work and once logged in the user will remain logged in
for 30 days unless they log out.

TO DO : Log Out functionality to be updated to remove access and refresh tokens
by getting API server to return cookies with expiry date in the past.