TIDY UP : ADD LOG OUT FUNCTIONALITY, CUSTOM MODAL, LOAD SPINNER, START TO CLEAN UP CODE
Added Log out functionality to destroy access and refresh tokens. Important
- also destroy local vuex state - otherwise coudl persist if new user logs
in straight after old one logs out.

Created a load spinner that is visible during server requests in case of
prolonged download

Started to clean up code by splitting long methods and functions
and keepign same level of abstraction. Also moved the contact API fetch
method into a general root action to prevent code duplication and keep
app DRY.

Created a custom model component that can be called when warning a user
prior to a DELETE API call. This was done using VUEX and creating a
Javascript promise object that is resolved when the user clicks 'yes'
or 'no' in the modal. The model html is teleported to the body to
function better and is only visible if the modal is required.

Put the code in local XAMPP environment - the Custom OOP PHP code
was placed in a sub folder /api - this means api calls are through the
same URI which means CORS issues are avoided.

TO DO: Add some custom CSS and VUE controlled animations & final
error check.